+++ front/src/App.js
@@ -5,7 +5,7 @@ import Login from "./component/login/Login";
 import SignUp from "./component/signUp/SignUp";
 import MainPage from "./component/mainPage/MainPage";
 import YearChange from "./component/yearChange/YearChange";
-import FindInfo from "./component/find_Info/FindID";
+import FindInfo from "./component/find_Info/FindMain";
 
 function App() {
   return (
@@ -15,6 +15,7 @@ function App() {
       <Route path="/mainPage" element={<MainPage />} />
       <Route path="/yearChange" element={<YearChange />} />
       <Route path="/find-id" element={<FindInfo />} />
+      <Route path="/find-password" element={<FindInfo />} />
     </Routes>
   );
 }
